requirements.txt
streamlit
openai
pinecone-client
BeautifulSoup4
tiktoken